requirements.txt
autopep8==2.0.4
flake8==7.0.0
pandas==2.2.1
mysql-connector==2.2.9